backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-socketio==5.11.2
motor==3.4.0
pymongo==4.7.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
python-dotenv==1.0.1
redis==5.0.4
celery==5.4.0
cloudinary==1.40.0
httpx==0.27.0
pydantic[email]==2.7.1
pydantic-settings==2.2.1
